Factor web app into modules
